refactor(server): llm logic adapte to rig lib
