+++ apps/server/api/src/ws/session/mod.rs
@@ -90,7 +90,7 @@ where
             .output_tx
             .clone()
             .expect("tx not create,maybe new round method before output frame method");
-        let llm = LlmFactory::global().get_llm();
+        let llm = LlmFactory::global().get_client();
         let tts = TtsCache::global().instance.clone();
         self.current_round = Some(Box::new(Round::new(
             self.id.clone(),
